requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27
channels[daphne]>=4.1
channels-redis>=4.2
redis>=5.0
python-dotenv>=1.0
